+++ lib/my_app.dart
@@ -1,0 +1,17 @@
+import 'package:flutter/material.dart';
+import 'package:flutter_guchi_memo/pages/guchi_home_page.dart';
+
+class MyApp extends StatelessWidget {
+  const MyApp({Key? key}) : super(key: key);
+  @override
+  Widget build(BuildContext context) {
+    return MaterialApp(
+      debugShowCheckedModeBanner: false,
+      title: 'GuchiRO',
+      theme: ThemeData(
+        primarySwatch: Colors.pink,
+      ),
+      home: const GuchiHomePage(),
+    );
+  }
+}
